reorganizacao de estrutura do projeto

+++ src/main/java/com/projetodaca/hitfire/repositories/LinkRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.projetodaca.hitfire.model.Link;
 
-public interface LinkRepository extends CrudRepository<Link, Long> {
+public interface LinkRepository extends CrudRepository<Link, Integer> {
 
 }
