+++ src/main/java/com/projetodaca/hitfire/repositories/MidiaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.projetodaca.hitfire.model.Midia;
 
-public interface MidiaRepository extends CrudRepository<Midia, Long> {
+public interface MidiaRepository extends CrudRepository<Midia, Integer> {
 
 }
